requirements.txt
hassapi~=0.1.1
appdaemon~=4.4.2

requests
pytest~=7.4.1
pytest-docker
pyyaml
appdaemon-testing